PYTHON/requirements.txt
customtkinter==5.2.2
pillow
pandas
scikit-learn
python-dotenv
google-generativeai
joblib
